+++ Assets/Mirror/Tests/Editor/Weaver/BatchedTestAttribute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b98ba1c237db0e94cbefc4a6517a5d5b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
