+++ Assets/Mirror/Tests/Editor/Weaver/BatchedWeaverTests.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -29,51 +28,7 @@ namespace Mirror.Weaver.Tests
 
     */
 
-    public class WeaverTestResults
-    {
-        public bool compileError;
-        public bool weaverError;
-        public List<string> weaverErrors = new List<string>();
-        public List<string> weaverWarnings = new List<string>();
 
-        public void CopyFrom(WeaverTestResults source)
-        {
-            compileError = source.compileError;
-            weaverError = source.weaverError;
-
-            weaverErrors.Clear();
-            weaverErrors.AddRange(source.weaverErrors);
-
-            weaverWarnings.Clear();
-            weaverWarnings.AddRange(source.weaverWarnings);
-        }
-
-        public void Clear()
-        {
-            compileError = false;
-            weaverError = false;
-            weaverErrors.Clear();
-            weaverWarnings.Clear();
-        }
-
-        public bool HasErrors()
-        {
-            return compileError
-                || weaverError
-                || weaverErrors.Count > 0;
-        }
-    }
-
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    public class BatchedTestAttribute : TestAttribute
-    {
-        public BatchedTestAttribute(bool shouldCompileWithNoErrors)
-        {
-            ShouldCompileWithNoErrors = shouldCompileWithNoErrors;
-        }
-
-        public bool ShouldCompileWithNoErrors { get; set; }
-    }
 
     /// <summary>
     /// Base class to used to run tests in baches
