+++ Assets/Mirror/Tests/Editor/Weaver/WeaverTestResults.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 746766023b2dfef4cae93bb13571d231
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
